Improved: variable name for preorder and backorder facility string and moved the logic to fetch the shipment detail in order service(#27xkwwd)
