+++ src/store/modules/order/actions.ts
@@ -56,7 +56,7 @@ const actions: ActionTree<OrderState, RootState> = {
         })
 
         if (completedOrderIds.length) {
-          orderItemTrackingCodes = await dispatch('fetchShipmentDetailForOrderItem', completedOrderIds)
+          orderItemTrackingCodes = await OrderService.getShipmentDetailForOrderItem(completedOrderIds)
         }
 
         const statuses = await this.dispatch('util/fetchStatus', [...status])
@@ -224,44 +224,6 @@ const actions: ActionTree<OrderState, RootState> = {
       console.error(err)
       showToast(translate('Something went wrong'))
     }
-  },
-
-  async fetchShipmentDetailForOrderItem({ commit }, payload) {
-    let resp;
-
-    // TODO: implement grouping logic to get the tracking code for order items
-    // currently when grouping only getting response for two groups thus used in operator for now
-    const params = {
-      "inputFields": {
-        "primaryOrderId": payload,
-        "primaryOrderId_op": "in",
-        "trackingCode_op": "not-empty",
-        "orderItemSeqId_op": "not-empty"
-      },
-      "entityName": "ShipmentPackageRouteSegAndItemDetail",
-      "distinct": "Y",
-      "noConditionFind": "Y",
-      "fieldList": ["trackingCode", "primaryOrderId", "orderItemSeqId"],
-      "viewSize": 100
-    }
-
-    try {
-      resp = await OrderService.fetchShipmentDetailForOrderItem(params)
-      if(resp.status == 200 && !hasError(resp) && resp.data.count > 0) {
-        return resp.data.docs.reduce((acc: any, list: any) => {
-          if(!acc[list.primaryOrderId]) {
-            acc[list.primaryOrderId] = {}
-          }
-          acc[list.primaryOrderId][list.orderItemSeqId] = list.trackingCode
-          return acc;
-        }, {})
-      } else {
-        console.error('Something went wrong while fetching tracking code for order items or there may not exist any tracking code for current order ids')
-      }
-    } catch(err) {
-      console.error(err)
-    }
-    return resp;
   }
 }
